+++ tests/phpunit/wpmatomo/test-paths.php
@@ -58,6 +58,10 @@ class PathsTest extends MatomoUnit_TestCase {
 	 * @group ms-required
 	 */
 	public function test_get_upload_base_dir_forBlog() {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Not multisite.' );
+			return;
+		}
 		$blogid1 = self::factory()->blog->create();
 		switch_to_blog( 2 );
 		$this->assertSame( get_temp_dir() . 'wordpress/wp-content/uploads/sites/2/matomo', $this->paths->get_upload_base_dir() );
@@ -183,10 +187,14 @@ class PathsTest extends MatomoUnit_TestCase {
 
 	public function test_clear_assets_dir_does_not_fail() {
 		$this->paths->clear_assets_dir();
+		// for the phpunit warning
+		$this->assertTrue( true );
 	}
 
 	public function test_clear_cache_dir_does_not_fail() {
 		$this->paths->clear_cache_dir();
+		// for the phpunit warning
+		$this->assertTrue( true );
 	}
 
 	public function test_get_relative_dir_to_matomo_differentDirectoryInPlugin() {
@@ -215,6 +223,10 @@ class PathsTest extends MatomoUnit_TestCase {
 	 * @group ms-required
 	 */
 	public function test_get_gloal_upload_dir_if_possible_forBlog() {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Not multisite.' );
+			return;
+		}
 		$blogid1 = self::factory()->blog->create();
 		switch_to_blog( 2 );
 		wp_delete_site( $blogid1 );
